rename and move attribute hooks
