+++ static/app/views/explore/hooks/useTraceItemAttributes.tsx
